integrate new uq cost model
